collect, partitions, run_length
