+++ src/main/scala/magic_rdds/collect.scala
@@ -1,0 +1,8 @@
+package magic_rdds
+
+import org.hammerlab.magic.rdd.collect.CollectPartitions
+
+trait collect
+  extends CollectPartitions
+
+case object collect extends collect
