+++ src/main/scala/magic_rdds/run_length.scala
@@ -1,0 +1,7 @@
+package magic_rdds
+
+import org.hammerlab.magic.rdd.RunLength
+
+trait run_length extends RunLength
+
+object run_length extends run_length
