+++ src/main/scala/org/hammerlab/magic/rdd/keyed/SplitByKeyRDD.scala
@@ -1,10 +1,9 @@
 package org.hammerlab.magic.rdd.keyed
 
+import magic_rdds.partitions._
 import org.apache.spark.rdd.RDD
 import org.hammerlab.iterator.CountIteratorByKey._
 import org.hammerlab.iterator.sliding.Sliding2Iterator._
-import org.hammerlab.magic.rdd.partitions.PartitionByKeyRDD._
-import org.hammerlab.magic.rdd.partitions.ReducePartitionsRDD._
 import org.hammerlab.magic.rdd.partitions.SlicePartitionsRDD
 import org.hammerlab.spark.PartitionIndex
