+++ src/main/scala/org/hammerlab/magic/rdd/rev.scala
@@ -1,7 +1,7 @@
 package org.hammerlab.magic.rdd
 
+import magic_rdds.partitions._
 import org.apache.spark.rdd.RDD
-import org.hammerlab.magic.rdd.partitions.PartitionByKeyRDD._
 
 import scala.reflect.ClassTag
