+++ src/main/scala/org/hammerlab/magic/rdd/sliding/SlidingRDD.scala
@@ -1,12 +1,12 @@
 package org.hammerlab.magic.rdd.sliding
 
+import magic_rdds.partitions._
 import magic_rdds.zip._
 import org.apache.spark.rdd.RDD
 import org.hammerlab.iterator.DropRightIterator._
 import org.hammerlab.iterator.sliding.SlidingIterator._
 import org.hammerlab.iterator.{ NextOptionIterator, SimpleBufferedIterator }
 import org.hammerlab.kryo._
-import org.hammerlab.magic.rdd.partitions.PartitionByKeyRDD._
 import org.hammerlab.spark.PartitionIndex
 
 import scala.reflect.ClassTag
